댓글 작성 폼

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -4,36 +4,35 @@
 <%@ include file="../common/head.jspf"%>
 
 <script>
-  const params = {};
-  params.id = parseInt('${param.id}');
+	const params = {};
+	params.id = parseInt('${param.id}');
 </script>
 
 <script>
-function ArticleDetail__increaseHitCout() {
-	const localStorageKey = 'article__'+ params.id + '__viewDone';
-	
-	if (localStorage.getItem(localStorageKey)) {
-		return;
-	}
-		
-	localStorage.setItem(localStorageKey, true);
-	
-	$.get('../article/doIncreaseHitCountRd', {
-			id: params.id,
-			ajaxMode: 'Y'
+	function ArticleDetail__increaseHitCout() {
+		const localStorageKey = 'article__' + params.id + '__viewDone';
+
+		if (localStorage.getItem(localStorageKey)) {
+			return;
+		}
+
+		localStorage.setItem(localStorageKey, true);
+
+		$.get('../article/doIncreaseHitCountRd', {
+			id : params.id,
+			ajaxMode : 'Y'
 		}, function(data) {
 			$('.article-detail__hit-count').empty().html(data.data1);
 		}, 'json');
-}
+	}
 
 	$(function() {
 		// 실전코드
 		// ArticleDetail__increaseHitCout();
-		
+
 		// 임시코드
 		setTimeout(ArticleDetail__increaseHitCout, 300)
 	})
-
 </script>
 
 <section class="mt-5">
@@ -72,37 +71,23 @@ function ArticleDetail__increaseHitCout() {
               <div class="flex items-center">
                 <span class="text-blue-700">${article.goodReactionPoint}</span>
                 <span>&nbsp;</span>
-                
+
                 <c:if test="${actorCanMakeReaction}">
-                  <a href="/usr/reactionPoint/doGoodReaction?relTypeCode=article&relId=${param.id}&replaceUri=${rq.encodedCurrentUri}" class="btn btn-xs  btn-primary btn-outline">
-                    좋아요
-                    👍
-                  </a>
+                  <a href="/usr/reactionPoint/doGoodReaction?relTypeCode=article&relId=${param.id}&replaceUri=${rq.encodedCurrentUri}" class="btn btn-xs  btn-primary btn-outline"> 좋아요 👍 </a>
                   <span>&nbsp;</span>
-                  <a href="/usr/reactionPoint/doBadReaction?relTypeCode=article&relId=${param.id}&replaceUri=${rq.encodedCurrentUri}" class="btn btn-xs  btn-secondary btn-outline">
-                    싫어요
-                    👎
-                  </a>
+                  <a href="/usr/reactionPoint/doBadReaction?relTypeCode=article&relId=${param.id}&replaceUri=${rq.encodedCurrentUri}" class="btn btn-xs  btn-secondary btn-outline"> 싫어요 👎 </a>
                 </c:if>
-                
+
                 <c:if test="${actorCanCancleGoodReaction}">
-                  <a href="/usr/reactionPoint/doCancleGoodReaction?relTypeCode=article&relId=${param.id}&replaceUri=${rq.encodedCurrentUri}" class="btn btn-xs  btn-primary">
-                    좋아요 👍
-                  </a>
+                  <a href="/usr/reactionPoint/doCancleGoodReaction?relTypeCode=article&relId=${param.id}&replaceUri=${rq.encodedCurrentUri}" class="btn btn-xs  btn-primary"> 좋아요 👍 </a>
                   <span>&nbsp;</span>
-                  <a onclick="alert(this.title); return false;" href="#" title="먼저 좋아요를 취소해주세요." class="btn btn-xs  btn-secondary btn-outline">
-                    싫어요 👎
-                  </a>
+                  <a onclick="alert(this.title); return false;" href="#" title="먼저 좋아요를 취소해주세요." class="btn btn-xs  btn-secondary btn-outline"> 싫어요 👎 </a>
                 </c:if>
-                
+
                 <c:if test="${actorCanCancleBadReaction}">
-                  <a onclick="alert(this.title); return false;" href="#" title="먼저 싫어요를 취소해주세요." class="btn btn-xs  btn-primary  btn-outline">
-                    좋아요 👍
-                  </a>
+                  <a onclick="alert(this.title); return false;" href="#" title="먼저 싫어요를 취소해주세요." class="btn btn-xs  btn-primary  btn-outline"> 좋아요 👍 </a>
                   <span>&nbsp;</span>
-                  <a href="/usr/reactionPoint/doCancleBadReaction?relTypeCode=article&relId=${param.id}&replaceUri=${rq.encodedCurrentUri}" class="btn btn-xs  btn-secondary">
-                    싫어요 👎
-                  </a>
+                  <a href="/usr/reactionPoint/doCancleBadReaction?relTypeCode=article&relId=${param.id}&replaceUri=${rq.encodedCurrentUri}" class="btn btn-xs  btn-secondary"> 싫어요 👎 </a>
                 </c:if>
               </div>
             </td>
@@ -121,14 +106,51 @@ function ArticleDetail__increaseHitCout() {
     <div class="btns">
       <button class="btn btn-link" type="button" onclick="history.back();">뒤로가기</button>
       <c:if test="${article.extra__actorCanModify}">
-        <a class="btn btn-link" " href="../article/modify?id=${article.id}">게시물 수정</a>
+        <a class="btn btn-link" href="../article/modify?id=${article.id}">게시물 수정</a>
       </c:if>
       <c:if test="${article.extra__actorCanDelete}">
-        <a class="btn btn-link" " onclick="if ( confirm('정말 삭제하시겠습니까?') == false) return false;"
-          href="../article/doDelete?id=${article.id}"
-        >게시물 삭제</a>
+        <a class="btn btn-link" onclick="if ( confirm('정말 삭제하시겠습니까?') == false) return false;" href="../article/doDelete?id=${article.id}">게시물 삭제</a>
       </c:if>
     </div>
+  </div>
+</section>
+
+<section class="mt-5">
+  <div class="container mx-auto px-3">
+    <h1>댓글 작성</h1>
+    <c:if test="${rq.logined}">
+      <form class="table-box-type-1" method="POST" action="../reply/doWrite">
+        <input type="hidden" name="relTypeCode" value="article" />
+        <input type="hidden" name="relId" value="${article.id}" />
+        <table>
+          <colgroup>
+            <col width="200" />
+          </colgroup>
+          <tbody>
+            <tr>
+            <tr>
+              <th>작성자</th>
+              <td>${rq.loginedMember.nickname}</td>
+            </tr>
+            <tr>
+              <th>내용</th>
+              <td>
+                <textarea required="required" name="body" rows="5" placeholder="내용"></textarea>
+              </td>
+            </tr>
+            <tr>
+              <th>댓글작성</th>
+              <td>
+                <button type="submit" class="btn btn-primary">댓글작성</button>
+              </td>
+            </tr>
+          </tbody>
+        </table>
+      </form>
+    </c:if>
+    <c:if test="${rq.notLogined}">
+      <a class="btn btn-link" href="/usr/member/login">로그인</a>후 이용해주세요.
+    </c:if>
   </div>
 </section>
